requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Admin==1.6.1
Flask-Caching==2.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.7
requests==2.31.0
feedparser==6.0.10
ics==0.7.2
schedule==1.2.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
